requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==11.3.0
requests==2.32.4
django-cors-headers==4.3.0
whitenoise==6.5.0